+++ Frends.SMTP.SendEmail/CHANGELOG.md
@@ -1,11 +1,19 @@
 # Changelog
+
+## [1.1.0] - 2023-11-29
+### Fixed
+- Fixed issue with the attachments can't be given as expression
+
 ## [1.0.3] - 2023-11-03
+### Added
 - Fix missing XML comments.
 
 ## [1.0.2] - 2023-11-03
+### Added
 - Fix bug #20 - Missing Frends tasks metadata file
 
 ## [1.0.1] - 2023-05-15
+### Fixed
 - Fix bug #14 - NullReferenceException when field with 'Expression' type is empty
 
 ## [1.0.0] - 2022-02-09
