+++ Frends.SMTP.SendEmail/Frends.Smtp.SendEmail/Definitions/AttachmentType.cs
@@ -1,0 +1,18 @@
+﻿namespace Frends.SMTP.SendEmail.Definitions
+{
+    /// <summary>
+    /// Enumeration for Attachment type.
+    /// </summary>
+    public enum AttachmentType
+    {
+        /// <summary>
+        /// Select this if the attachment is a file.
+        /// </summary>
+        FileAttachment,
+
+        /// <summary>
+        /// Select this if the attachment file should be created from a string.
+        /// </summary>
+        AttachmentFromString
+    }
+}
