+++ Frends.SMTP.SendEmail/Frends.Smtp.SendEmail/Frends.Smtp.SendEmail.csproj
@@ -1,9 +1,9 @@
-
+﻿
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
   <TargetFrameworks>netstandard2.0;net6.0</TargetFrameworks>
-	<Version>1.0.3</Version>
+	<Version>1.0.4</Version>
 	<Authors>Frends</Authors>
 	<Company>Frends</Company>
 	<Product>Frends</Product>
