+++ Frends.SMTP.SendEmail/Frends.Smtp.SendEmail/GlobalSuppressions.cs
@@ -1,0 +1,4 @@
+﻿using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Following Frends guidelines", Scope = "type", Target = "~T:Frends.SMTP.SendEmail.SMTP")]
+[assembly: SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1633:File should have header", Justification = "Following Frends guidelines")]
