+++ Frends.SMTP.SendEmail/README.md
@@ -1,11 +1,10 @@
 # Frends.SMTP.SendEmail
-
-
-[![Frends.Smtp.SendEmail Main](https://github.com/FrendsPlatform/Frends.Smtp/actions/workflows/SendEmail_main.yml/badge.svg)](https://github.com/FrendsPlatform/Frends.Regex/actions/workflows/IsMatch_build_and_test_on_main.yml)
- ![GitHub](https://img.shields.io/github/license/FrendsPlatform/Frends.Smtp?label=License)
- ![Coverage](https://app-github-custom-badges.azurewebsites.net/Badge?key=FrendsPlatform/Frends.Smtp/Frends.Smtp.SendEmail|main)
-
 Frends task for sending emails with SMTP. Task sends emails via SMTP protocol and can handle attachments either from file or as raw string input.
+
+
+[![Frends.Smtp.SendEmail Main](https://github.com/FrendsPlatform/Frends.SMTP/actions/workflows/SendEmail_main.yml/badge.svg)](https://github.com/FrendsPlatform/Frends.SMTP/actions)
+ ![GitHub](https://img.shields.io/github/license/FrendsPlatform/Frends.SMTP?label=License)
+ ![Coverage](https://app-github-custom-badges.azurewebsites.net/Badge?key=FrendsPlatform/Frends.SMTP/Frends.SMTP.SendEmail|main)
 
 ## Installing
 
@@ -13,14 +12,18 @@ You can install the Task via frends UI Task View.
 
 ## Building
 
-Rebuild the project
+### Clone a copy of the repository
+
+`git clone https://github.com/FrendsPlatform/Frends.SMTP.git`
+
+### Rebuild the project
 
 `dotnet build`
 
-Run tests
+### Run tests
 
 `dotnet test`
 
-Create a NuGet package
+### Create a NuGet package
 
 `dotnet pack --configuration Release`
