+++ README.md
@@ -4,7 +4,7 @@ Frends tasks for SMTP protocol.
 
 # Tasks
 
-- [Frends.Smtp.SendEmail](Frends.Smtp.SendEmail/README.md)
+- [Frends.SMTP.SendEmail](Frends.SMTP.SendEmail/README.md)
 
 # Contributing
 When contributing to this repository, please first discuss the change you wish to make via issue, email, or any other method with the owners of this repository before making a change.
